requirements.txt
scapy==2.4.5
tensorflow==2.12.0
pyyaml==6.0
numpy==1.25.0
pandas==2.2.0
scikit-learn==1.3.0
